docker/api/requirements.txt
# Local Agentic AI System — Python dependencies
# All packages are open-source (MIT/Apache/PSF licensed)
# Install: pip install -r requirements.txt

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9       # file upload support

# Configuration
pydantic-settings==2.3.4

# PostgreSQL 18.1 driver (psycopg3, open-source)
psycopg[binary]==3.1.19

# HTTP client (for calling Ollama)
httpx==0.27.0

# File extraction — open-source only
pdfplumber==0.11.0            # PDF text extraction
python-docx==1.1.2            # DOCX
python-pptx==0.6.23           # PPTX
openpyxl==3.1.2               # XLSX
Pillow==10.3.0                # Image processing
pytesseract==0.3.13           # OCR (requires Tesseract binary: apt install tesseract-ocr)
